Rename `igls.py` -> `ig_lightstreamer.py`.
